changed size_t to unsigned; included a few (yet unimplemented) methods for mesh-consistency checking

+++ MeshLib/Elements/Cell.cpp
@@ -9,12 +9,12 @@
 
 namespace MeshLib {
 /*
-Cell::Cell(Node** nodes, MshElemType::type type, size_t value)
+Cell::Cell(Node** nodes, MshElemType::type type, unsigned value)
 	: Element(nodes, type, value)
 {
 }
 */
-Cell::Cell(MshElemType::type type, size_t value)
+Cell::Cell(MshElemType::type type, unsigned value)
 	: Element(type, value)
 {
 }
