+++ MeshLib/Elements/Face.h
@@ -22,10 +22,10 @@ public:
 	virtual double getArea() const { return _area; };
 
 	/// Get dimension of the mesh element.
-	size_t getDimension() const { return 2; };
+	unsigned getDimension() const { return 2; };
 
 	/// 2D elements have no faces.
-	size_t getNFaces() const { return 0; };
+	unsigned getNFaces() const { return 0; };
 
 	/// Destructor
 	virtual ~Face();
@@ -33,10 +33,10 @@ public:
 protected:
 /*
 	/// Constructor for a generic mesh element containing an array of mesh nodes.
-	Face(Node** nodes, MshElemType::type type, size_t value = 0);
+	Face(Node** nodes, MshElemType::type type, unsigned value = 0);
 */
 	/// Constructor for a generic mesh element without an array of mesh nodes.
-	Face(MshElemType::type type, size_t value = 0);
+	Face(MshElemType::type type, unsigned value = 0);
 
 	/// Calculate the area of this 2d element.
 	virtual double calcArea() = 0;
